refactor parse module into submodules
