+++ src/parse/mod.rs
@@ -1,5 +1,9 @@
-mod parse_pattern;
 mod token;
+mod utils;
+mod leaf;
+mod meta;
+mod structure;
+mod parse_pattern;
 
 pub use parse_pattern::parse_pattern;
 pub use token::Token;
