+++ src/parse/structure/mod.rs
@@ -1,0 +1,4 @@
+// Parsers for structure-level pattern syntax
+
+// Currently unused but reserved for future expansion.
+
